rename promAPI package to prom in monitoring integration test
