+++ tests/integration/monitoring/prom/targets.go
@@ -1,4 +1,4 @@
-package promAPI
+package prom
 
 // To parse JSON response from http://monitoring-prometheus.kyma-system:9090/api/v1/targets
 type TargetsResponse struct {
